Fix Table segments column avg_heartrate must be double precision not float. Closes #159.

+++ backend/database/init/02_create_tables.sql
@@ -61,7 +61,7 @@ CREATE TABLE segments (
     end_time INTEGER,
     start_heartrate INTEGER,
     end_heartrate INTEGER,
-    avg_heartrate INTEGER,
+    avg_heartrate DOUBLE PRECISION,
     avg_speed DOUBLE PRECISION,
     elevation_gain INTEGER,
     hr_drift DOUBLE PRECISION,
